refactor: rename mod workspaces -> workspace
