+++ libs/jwst/src/workspace/block_observer.rs
