+++ libs/jwst/src/workspace/metadata/mod.rs
