+++ libs/jwst/src/workspace/plugins/indexing/indexer.rs
