+++ libs/jwst/src/workspace/plugins/indexing/mod.rs
