+++ libs/jwst/src/workspace/plugins/indexing/register.rs
