+++ libs/jwst/src/workspace/plugins/plugin.rs
