+++ libs/jwst/src/workspace/workspace.rs
@@ -10,6 +10,7 @@ use yrs::{
     types::{map::MapEvent, ToJson},
     Doc, Map, MapRef, Subscription, Transact, TransactionMut, UpdateSubscription,
 };
+use crate::workspace::block_observer::BlockObserverConfig;
 
 pub type MapSubscription = Subscription<Arc<dyn Fn(&TransactionMut, &MapEvent)>>;
